Face_Emotion-main/Face/requirements.txt
numpy
streamlit
keras
opencv-contrib-python-headless
tensorflow-cpu
streamlit_webrtc
